Basic epub generation
This is a work in progress. Right now the generate_epub method uses the
markdown source files of each chapter and feeds them to Pandoc.

This works on a crude level, and generates a workable TOC.
The problems with this approach are:

- Kramdown extensions get left in the text
- Any data in the chapter frontmatter is left out of the ebook

As an attempt to solve the second problem here, a basic epub_layout has
been added but it is not currently being used.

One potential solution might be to try and use the ruby_pandoc gem and
feed data to pandoc from memory instead of via the shell (this would
maybe avoid any need to generate temporary files and clean up).