WIP: Epub generation
This is a work in progress. The current commit adds a number of modules
to automate epub generation. The heavy lifting in terms of XML parsing
and writing comes from Nokogiri, which is already used by Middleman for
other things.

The basic process is that an instance of Book::Epub gets created if the
-epub flag is sent at build time. This class has a number of methods
which are essentially nokogiri scripts.

An accompanying folder of templates lives inside the extension as
well (and these will be included in any future gem release). These are
parsed by Nokogiri and used as the raw material for the epub HTML and
XML components.

For the content of the book, Chapter resource objects are passed in to
the book instance at initialization time and the string contents of
these chapters (created by the #render method) is fed into Nokogiri as a
DocumentFragment. Then this is appended to the appropriate location in
the template and the resulting document is written to a file.

Ultimately, this module will write a complete, expanded Epub to a folder
in the dist/ directory which can then be zipped as an epub file.
Validation should happen at the end of this process as well.